demTest/requirements.txt
affine==2.4.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
fonttools==4.60.1
freetype-py==2.5.1
glfw==2.10.0
hsluv==5.0.4
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.8.0
numpy==1.26.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
pyglet==2.1.9
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
pyparsing==3.2.5
PyQt5-Qt5==5.15.2
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
pyqt6-tools==6.4.2.3.3
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
QtPy==2.4.3
rasterio==1.4.3
requests==2.32.5
scipy==1.13.1
scooby==0.10.2
shapely==2.0.7
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
zipp==3.23.0
